requirements.txt
kafka-python==1.4.6
psycopg2-binary==2.8.2
requests==2.22.0